requirements.txt
PyQt6==6.6.0
qrcode==7.4.2
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
matplotlib==3.8.2


